Move icinga2-classicui to its own configuration directory.
/etc/icinga2 may contain sensitive information like passwords or
the salt hash for the ticket generator. Therefore we decided to
remove read permissions for non-nagios users. That means that the
classic-ui can't read its configuration files anymore.

We moved those files into its own directory /etc/icinga2-classicui.
So if you are looking for classicui related configurations, this is
the place too look